+++ packages/data-curation/src/data_curation/terminologies/http_client.py
